+++ src/state_machines/PiquetStates/piquet_idle_state.cpp
@@ -33,7 +33,7 @@ void PiquetIdleState::onExecute()
         return;
     }
 
-    unsigned long currentMillis = millis();
+    uint32_t currentMillis = millis();
     if (currentMillis - previousMillis >= PIQUET_SPEED)
     {
         previousMillis = currentMillis;
