+++ src/state_machines/PiquetStates/piquet_idle_state.h
@@ -20,7 +20,7 @@ public:
 
 private:
     PiquetController *_controller;
-    unsigned long previousMillis = 0;
+    uint32_t previousMillis = 0;
 };
 
 #endif
